adding pagination count

+++ src/main/java/com/covimyn/search/dao/ResourceDao.java
@@ -17,5 +17,5 @@ public interface ResourceDao {
     String upsert(ResourceModel resourceModel) throws IOException;
     List<ResourceModel> search(List<Pair> must, List<Pair> should, int page, int size) throws Exception;
     List<ResourceModel> searchByLatest(List<Pair> must, List<Pair> should, int page, int size) throws Exception;
-    int count(List<Pair> must, List<Pair> should) throws Exception;
+    long count(List<Pair> must, List<Pair> should) throws Exception;
 }
